luxury-agency-backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-filter==24.3
Pillow==10.4.0
